Refactoring the feature spec + page spec objects across the board
