+++ spec/pages/admin/patients/appointments_page.rb
@@ -1,0 +1,2 @@
+class AppointmentsPage < ApplicationPage
+end
