+++ spec/pages/admin/patients/patients_page.rb
@@ -1,0 +1,2 @@
+class PatientsPage < ApplicationPage
+end
