--- spec/pages/dashboard/dashboard_page.rb
@@ -1,8 +1,0 @@
-class DashboardPage < ApplicationPage
-
-  ORGANIZATIONS = {css: "div.card" }.freeze
-
-  def get_organization_count
-    all_elements(ORGANIZATIONS).size
-  end
-end
