--- spec/pages/facility_groups/admin_facility_group_page_new.rb
@@ -1,58 +1,0 @@
-class AdminFacilityGroupPageNew < ApplicationPage
-
-  CREATE_FACILITY_GROUP_BUTTON = { css: "input[value='Create Facility group']"}.freeze
-  FACILITY_NAME = { id: 'facility_name' }.freeze
-  FACILITY_DESCRIPTION = { id: 'facility_description' }.freeze
-  DELETE_FACILITY_GROUP_BUTTON = {css: "a.ml-4"}.freeze
-  PROTOCOL_DROPDOWN = { xpath: "//select[@name='facility_group[protocol_id]']" }.freeze
-  UNASSOCIATED_FACILITY_CHECKBOX = { css: "input[type='checkbox']" }.freeze
-  SUCCESSFUL_MESSAGE = { css: 'div.alert-primary'}.freeze
-  MESSAGE_CROSS_BUTTON = {css: "button.close" }.freeze
-  UPDATE_FACILITY_GROUP_BUTTON = { css: "input[value='Update Facility group']"}.freeze
-
-
-  def select_organisation_name_dropdown(value)
-    find(:xpath, "//select[@name='facility_group[organization_id]']").find(:option, value).select_option
-  end
-
-  def select_protocol_name_dropdown(value)
-    find(:xpath, "//select[@name='facility_group[protocol_id]']").find(:option, value).select_option
-  end
-
-  def add_new_facility_group_without_assigningfacility(org_name, name, description, protocol_name)
-
-    select_organisation_name_dropdown(org_name)
-    type(FACILITY_NAME, name)
-    type(FACILITY_DESCRIPTION, description)
-    select_protocol_name_dropdown(protocol_name)
-    click(CREATE_FACILITY_GROUP_BUTTON)
-  end
-
-  def add_new_facility_group(org_name, name, description, unassociatedfacility, protocol_name)
-
-    select_organisation_name_dropdown(org_name)
-    type(FACILITY_NAME, name)
-    type(FACILITY_DESCRIPTION, description)
-    select_unassociated_facility(unassociatedfacility)
-    select_protocol_name_dropdown(protocol_name)
-    click(CREATE_FACILITY_GROUP_BUTTON)
-  end
-
-  def click_on_delete_facility_group_button
-    click(DELETE_FACILITY_GROUP_BUTTON)
-    # page.accept_alert("OK")
-    # present?(SUCCESSFUL_MESSAGE)
-    # click(MESSAGE_CROSS_BUTTON)
-  end
-
-  def select_unassociated_facility(facility_name)
-    within(:xpath ,"//label[text()='Unassociated facilities']/..")do
-      page.has_content?(facility_name)
-      find(:xpath, "//div[@class='form-check']/label[text()='#{facility_name}']/../input").click
-    end
-  end
-
-  def click_on_update_facility_group_button
-    click(UPDATE_FACILITY_GROUP_BUTTON)
-  end
-end
