+++ spec/pages/navigations/dashboard_page_navigation.rb
@@ -1,5 +1,4 @@
 class DashboardPageNavigation < ApplicationPage
-
   LOGOUT_BUTTON = { css: 'a.dropdown-item[href="/admins/sign_out"]'}.freeze
   MANAGE_OPTION = { xpath: "//li/div/a"}.freeze
   MAIN_MENU_TABS = { css: "ul.mr-auto>li>a" }.freeze
