+++ spec/pages/organizations/index.rb
@@ -1,0 +1,9 @@
+module OrganizationsPage
+  class Index < ApplicationPage
+    ORGANIZATIONS = { css: "div.card" }.freeze
+
+    def get_organization_count
+      all_elements(ORGANIZATIONS).size
+    end
+  end
+end
