--- spec/pages/patients/appointments_page.rb
@@ -1,3 +1,0 @@
-class AppointmentsPage < ApplicationPage
-
-end
