--- spec/pages/patients/patients_page.rb
@@ -1,4 +1,0 @@
-class PatientsPage < ApplicationPage
-
-
-end
